toolkit/devtools/server/tests/mochitest/test_inspector-reload.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
Components.utils.import("resource://gre/modules/devtools/Loader.jsm");
const {Promise: promise} = Components.utils.import("resource://gre/modules/Promise.jsm", {});

const inspector = devtools.require("devtools/server/actors/inspector");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gInspectee = null;
var gClient = null;
var gWalker = null;

addTest(function setup() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    let {InspectorFront} = devtools.require("devtools/server/actors/inspector");
    let inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
      return inspector.getWalker();
    }).then(walker => {
      dump(walker.actorID + "\n");
      ok(walker === gWalker, "getWalker() twice should return the same walker.");
    }).then(runNextTest));
  });
});

addTest(function testReload() {
  let nodeFront;
  let oldRootID = gWalker.rootNode.actorID;
  // Load a node to populate the tree a bit.
  promiseDone(gWalker.querySelector(gWalker.rootNode, "#a").then(front => {
    gInspectee.defaultView.location.reload();
    return waitForMutation(gWalker, isNewRoot);
  }).then(() => {
    ok(gWalker.rootNode.actorID != oldRootID, "Root node should have changed.");
  }).then(() => {
    // Make sure we can still access the document
    return gWalker.querySelector(gWalker.rootNode, "#a");
  }).then(front => {
    ok(front.actorID, "Got a new actor ID");
  }).then(runNextTest));
});

addTest(function cleanup() {
  delete gWalker;
  delete gInspectee;
  delete gClient;
  runNextTest();
});


  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
